Turned PdbBlockDervier's component discovery code into a separate service.
