+++ pdb.services.yml
@@ -1,0 +1,7 @@
+services:
+  pdb.component_discovery:
+    class: '\Drupal\pdb\ComponentDiscovery'
+    arguments:
+      - '@app.root'
+      - '@module_handler'
+      - '@info_parser'
