+++ src/ComponentDiscoveryInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * @file
+ * Contains \Drupal\pdb\ComponentDiscoveryInterface.
+ */
+
+namespace Drupal\pdb;
+
+interface ComponentDiscoveryInterface {
+
+  public function getComponents();
+
+}
